Fix "stuck buttons" that would occur with the ButtonMouse mode.
Add a toggle for ButtonMouse mode which lets you have buttons
enabled without having a finger down.  Press down on the touchpad
with two fingers to toggle between hybrid controller mode and the
pure mouse mode which always consumes all mouse-button-equivalent
button presses regardless of your finger being on the touchpad.
This still reserves the touchpad single-finger-push gesture for
a custom keybind, but it removes the ability to rebind the
double-finger-push just for this particular mouse mode.  This mode
should make it very easy to click+drag, double-click, etc.