Add haptic feedback for the ButtonMouse mode.
Add three possible sections for the touchpad as keybinding on
the ButtonMouse: the Left section (configured with "Touchpad"),
Right section (configured with "Multitouch") and then the middle
and top bar section are configured by "Upperpad".

Switching between "Button Lock" (two finger push) in ButtonMouse
on and off have haptic feedback to tell you which mode you are
going into.